【问题/需求】迁移SiameseNetwork与MiniCPM清理改动 【原因分析】将本地modelzoo-dev中的样例文档、构建配置及前后处理代码同步到提交仓库 【修改描述】同步SiameseNetwork README、MiniCPM main.cpp、build_config以及common前后处理实现 【注意事项】不涉及 【涉及芯片】Hi3403V100 【自测用例】使用modelzoo-submit脚本迁移指定6个文件并执行git提交 【自测结果】脚本完成文件同步并生成提交

+++ samples/built-in/vlm/MiniCPM/src/main.cpp
@@ -30,7 +30,7 @@ int main(int argc, char* argv[])
     if (param.loop == ZERO)
         param.loop++;
     string result;
-    for (int i = 0; i < param.loop; ++i) {
+    for (size_t i = 0; i < param.loop; ++i) {
         for (const auto& prompt : param.fileLists) {
             if (prompt.size() != QUESTION_PAIR_ITEM_NUM) {
                 LOG(ERROR) << "Invalid file_list.json format: questions must be defined in image-question pairs";
